// src/generated/wotlk/ui.ts
export interface UIItem {
    /**
     * @generated from protobuf field: int32 id = 1
     */
    id: number;
    /**
     * @generated from protobuf field: string name = 2
     */
    name: string;
    /**
     * @generated from protobuf field: string icon = 3
     */
    icon: string;
    /**
     * @generated from protobuf field: proto.ItemType type = 4
     */
    type: ItemType;
    /**
     * @generated from protobuf field: proto.ArmorType armor_type = 5
     */
    armorType: ArmorType;
    /**
     * @generated from protobuf field: proto.WeaponType weapon_type = 6
     */
    weaponType: WeaponType;
    /**
     * @generated from protobuf field: proto.HandType hand_type = 7
     */
    handType: HandType;
    /**
     * @generated from protobuf field: proto.RangedWeaponType ranged_weapon_type = 8
     */
    rangedWeaponType: RangedWeaponType;
    /**
     * @generated from protobuf field: repeated double stats = 9
     */
    stats: number[];
    /**
     * @generated from protobuf field: repeated proto.GemColor gem_sockets = 10
     */
    gemSockets: GemColor[];
    /**
     * @generated from protobuf field: repeated double socketBonus = 11
     */
    socketBonus: number[];
    /**
     * Weapon stats, needed for computing proper EP for melee weapons
     *
     * @generated from protobuf field: double weapon_damage_min = 12
     */
    weaponDamageMin: number;
    /**
     * @generated from protobuf field: double weapon_damage_max = 13
     */
    weaponDamageMax: number;
    /**
     * @generated from protobuf field: double weapon_speed = 14
     */
    weaponSpeed: number;
    /**
     * @generated from protobuf field: int32 ilvl = 15
     */
    ilvl: number;
    /**
     * @generated from protobuf field: int32 phase = 16
     */
    phase: number;
    /**
     * @generated from protobuf field: proto.ItemQuality quality = 17
     */
    quality: ItemQuality;
    /**
     * @generated from protobuf field: bool unique = 18
     */
    unique: boolean;
    /**
     * @generated from protobuf field: bool heroic = 19
     */
    heroic: boolean;
    /**
     * Classes that are allowed to use the item. Empty indicates no special class restrictions.
     *
     * @generated from protobuf field: repeated proto.Class class_allowlist = 20
     */
    classAllowlist: Class[];
    /**
     * @generated from protobuf field: proto.Profession required_profession = 21
     */
    requiredProfession: Profession;
    /**
     * Name of the item set to which this item belongs, if any.
     *
     * @generated from protobuf field: string set_name = 22
     */
    setName: string;
    /**
     * @generated from protobuf field: proto.Expansion expansion = 24
     */
    expansion: Expansion;
    /**
     * @generated from protobuf field: repeated proto.UIItemSource sources = 23
     */
    sources: UIItemSource[];
    /**
     * @generated from protobuf field: proto.UIItem.FactionRestriction faction_restriction = 25
     */
    factionRestriction: UIItem_FactionRestriction;
}
/**
 * @generated from protobuf enum proto.UIItem.FactionRestriction
 */
export enum UIItem_FactionRestriction {
    /**
     * @generated from protobuf enum value: FACTION_RESTRICTION_UNSPECIFIED = 0;
     */
    UNSPECIFIED = 0,
    /**
     * @generated from protobuf enum value: FACTION_RESTRICTION_ALLIANCE_ONLY = 1;
     */
    ALLIANCE_ONLY = 1,
    /**
     * @generated from protobuf enum value: FACTION_RESTRICTION_HORDE_ONLY = 2;
     */
    HORDE_ONLY = 2
}
/**
 * @generated from protobuf message proto.UIItemSource
 */
export interface UIItemSource {
    /**
     * @generated from protobuf oneof: source
     */
    source: {
        oneofKind: "crafted";
        /**
         * @generated from protobuf field: proto.CraftedSource crafted = 1
         */
        crafted: CraftedSource;
    } | {
        oneofKind: "drop";
        /**
         * @generated from protobuf field: proto.DropSource drop = 2
         */
        drop: DropSource;
    } | {
        oneofKind: "quest";
        /**
         * @generated from protobuf field: proto.QuestSource quest = 3
         */
        quest: QuestSource;
    } | {
        oneofKind: "soldBy";
        /**
         * @generated from protobuf field: proto.SoldBySource sold_by = 4
         */
        soldBy: SoldBySource;
    } | {
        oneofKind: "rep";
        /**
         * @generated from protobuf field: proto.RepSource rep = 5
         */
        rep: RepSource;
    } | {
        oneofKind: undefined;
    };
}
/**
 * @generated from protobuf message proto.CraftedSource
 */
export interface CraftedSource {
    /**
     * @generated from protobuf field: proto.Profession profession = 1
     */
    profession: Profession;
    /**
     * @generated from protobuf field: int32 spell_id = 2
     */
    spellId: number;
}
/**
 * @generated from protobuf message proto.DropSource
 */
export interface DropSource {
    /**
     * @generated from protobuf field: proto.DungeonDifficulty difficulty = 1
     */
    difficulty: DungeonDifficulty;
    /**
     * @generated from protobuf field: int32 npc_id = 2
     */
    npcId: number;
    /**
     * @generated from protobuf field: int32 zone_id = 3
     */
    zoneId: number;
    /**
     * @generated from protobuf field: string other_name = 4
     */
    otherName: string; // For drops in zones from things other than NPCs, e.g. "Trash"
    /**
     * @generated from protobuf field: string category = 5
     */
    category: string; // For conditions/categories of drops, e.g. 'Hard Mode' or 'Titan Rune'
}
/**
 * @generated from protobuf message proto.QuestSource
 */
export interface QuestSource {
    /**
     * @generated from protobuf field: int32 id = 1
     */
    id: number;
    /**
     * @generated from protobuf field: string name = 2
     */
    name: string;
}
/**
 * @generated from protobuf message proto.SoldBySource
 */
export interface SoldBySource {
    /**
     * @generated from protobuf field: int32 npc_id = 1
     */
    npcId: number;
    /**
     * @generated from protobuf field: string npc_name = 2
     */
    npcName: string;
    /**
     * @generated from protobuf field: int32 zone_id = 3
     */
    zoneId: number;
}
/**
 * @generated from protobuf message proto.RepSource
 */
export interface RepSource {
    /**
     * @generated from protobuf field: proto.RepFaction rep_faction_id = 1
     */
    repFactionId: RepFaction;
    /**
     * @generated from protobuf field: proto.RepLevel rep_level = 2
     */
    repLevel: RepLevel;
    /**
     * @generated from protobuf field: proto.Faction faction_id = 3
     */
    factionId: Faction;
}
/**
 * @generated from protobuf message proto.UIEnchant
 */
export interface UIEnchant {
    /**
     * All enchants have an effect ID. Some also have an item ID, others have a spell ID,
     * and yet others have both item and spell IDs. No single ID alone is enough to
     * uniquely identify an enchant. Uniqueness requires either effect ID + slot, or
     * effect ID + item/spell ID.
     *
     * @generated from protobuf field: int32 effect_id = 1
     */
    effectId: number; // ID of the effect.
    /**
     * @generated from protobuf field: int32 item_id = 2
     */
    itemId: number; // ID of the enchant "item". Might be 0 if not available.
    /**
     * @generated from protobuf field: int32 spell_id = 3
     */
    spellId: number; // ID of the enchant "spell". Might be 0 if not available.
    /**
     * @generated from protobuf field: string name = 4
     */
    name: string;
    /**
     * @generated from protobuf field: string icon = 5
     */
    icon: string;
    /**
     * @generated from protobuf field: proto.ItemType type = 6
     */
    type: ItemType; // Which type of item this enchant can be applied to.
    /**
     * @generated from protobuf field: repeated proto.ItemType extra_types = 13
     */
    extraTypes: ItemType[]; // Extra types for enchants that can go in multiple slots (like armor kits).
    /**
     * @generated from protobuf field: proto.EnchantType enchant_type = 7
     */
    enchantType: EnchantType;
    /**
     * @generated from protobuf field: repeated double stats = 8
     */
    stats: number[];
    /**
     * @generated from protobuf field: proto.ItemQuality quality = 9
     */
    quality: ItemQuality;
    /**
     * @generated from protobuf field: int32 phase = 10
     */
    phase: number;
    /**
     * Classes that are allowed to use the enchant. Empty indicates no special class restrictions.
     *
     * @generated from protobuf field: repeated proto.Class class_allowlist = 11
     */
    classAllowlist: Class[];
    /**
     * @generated from protobuf field: proto.Profession required_profession = 12
     */
    requiredProfession: Profession;
}
// ...
export enum Expansion {
    /**
     * @generated from protobuf enum value: ExpansionUnknown = 0;
     */
    ExpansionUnknown = 0,
    /**
     * @generated from protobuf enum value: ExpansionVanilla = 1;
     */
    ExpansionVanilla = 1,
    /**
     * @generated from protobuf enum value: ExpansionTbc = 2;
     */
    ExpansionTbc = 2,
    /**
     * @generated from protobuf enum value: ExpansionWotlk = 3;
     */
    ExpansionWotlk = 3
}
// ...
export const UIItem = new UIItem$Type();
// ...
export const UIItemSource = new UIItemSource$Type();
// ...
export const CraftedSource = new CraftedSource$Type();
// ...
export const DropSource = new DropSource$Type();
// ...
export const QuestSource = new QuestSource$Type();
// ...
export const SoldBySource = new SoldBySource$Type();
// ...
export const RepSource = new RepSource$Type();
// ...
export const UIEnchant = new UIEnchant$Type();